Make individual test files simpler
This encodes them as some kind of DSL that is hopefully going to be
easier to extract at a later date.

The MMU1 test still needs some work: how do we detect what kind of test
it is? TLB invalidation is expensive to do on each iteration but might
be required for the MMU1 test... unless we make each heap variable take
an entire page!